Ajust new concept for habit
1. Adjust new concept for habit

2. New Editor layout

+++ app/src/main/java/com/dd/android/dailysimple/daily/edit/observable/AlarmObservable.kt
@@ -36,7 +36,7 @@ class AlarmObservable(alarm: Alarm) : BaseObservable() {
         }
 
     @get:Bindable
-    var alarmTime: Int
+    var alarmTime: Long
         get() = alarm.alarmTime
         set(value) {
             alarm.alarmTime = value
